Move Parser code to FloorSubsytem class.
